requirements.txt
aiolimiter==1.1.0
anyio==3.7.1
APScheduler==3.10.4
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
filelock==3.12.2
fsspec==2023.6.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
numpy==1.25.2
packaging==23.1
python-dotenv==0.15.0
python-telegram-bot==20.4
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
six==1.16.0
sniffio==1.3.0
tokenizers==0.13.3
tornado==6.3.3
tqdm==4.66.1
transformers==4.32.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
